working an exercise outside tutorial
