+++ ExampleComp
@@ -1,0 +1,13 @@
+import React, { Component} from 'react'
+
+class ExampleComp extends Component {
+    render() {
+        return() (
+            <div>This is a component</div>
+        
+        }
+    
+
+}
+
+export default ExampleComp
