+++ src/firstChild.jsx
@@ -1,0 +1,7 @@
+import React from 'react';     // React needs to be imported
+
+const FirstChild = () => {
+  return <p>I'm the 1st child!</p>; 
+};
+
+export default FirstChild;
